# api/swagger.yaml
basePath: /
definitions:
  caspercloud_internal_repository.Image:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      source_url:
        type: string
    type: object
...
  caspercloud_internal_repository.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
  caspercloud_internal_service.AuthResponse:
    properties:
      active_project_id:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/caspercloud_internal_repository.User'
    type: object
...
  internal_httpapi.docAuthData:
    properties:
      data:
        $ref: '#/definitions/caspercloud_internal_service.AuthResponse'
    type: object
...
  internal_httpapi.docImageData:
    properties:
      data:
        $ref: '#/definitions/caspercloud_internal_repository.Image'
    type: object
  internal_httpapi.docImagesData:
    properties:
      data:
        items:
          $ref: '#/definitions/caspercloud_internal_repository.Image'
        type: array
    type: object
...
  internal_httpapi.imageRequest:
    properties:
      description:
        maxLength: 4096
        type: string
      name:
        maxLength: 256
        minLength: 1
        type: string
      source_url:
        maxLength: 2048
        minLength: 1
        type: string
    required:
    - name
    - source_url
    type: object
  internal_httpapi.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      project_id:
        type: string
    required:
    - email
    - password
    type: object
  internal_httpapi.registerRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_httpapi.switchProjectRequest:
    properties:
      project_id:
        type: string
    required:
    - project_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'HTTP API for CasperCloud (multi-tenant projects, images, instances).
    Authenticated routes expect `Authorization: Bearer <JWT>`. Project-scoped routes
    require a token whose `active_project_id` matches the path `projectID` (use login
    with `project_id` or POST /v1/auth/switch-project).'
  title: CasperCloud API
  version: "1.0"
paths:
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docAuthData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_httpapi.docAuthData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register
      tags:
      - auth
  /v1/auth/switch-project:
    post:
      consumes:
      - application/json
      description: Returns a new token whose active_project_id matches the requested
        project (must be a member).
      parameters:
      - description: Target project
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.switchProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docAuthData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Switch active project
      tags:
      - auth
...
  /v1/projects/{projectID}/images:
    get:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docImagesData'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List images
      tags:
      - images
    post:
      consumes:
      - application/json
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Image metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.imageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_httpapi.docImageData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create image
      tags:
      - images
  /v1/projects/{projectID}/images/{imageID}:
    delete:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Image UUID
        in: path
        name: imageID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete image
      tags:
      - images
    get:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Image UUID
        in: path
        name: imageID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docImageData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get image
      tags:
      - images
    put:
      consumes:
      - application/json
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Image UUID
        in: path
        name: imageID
        required: true
        type: string
      - description: Image metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.imageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docImageData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update image
      tags:
      - images
...
securityDefinitions:
  BearerAuth:
    description: JWT from POST /v1/auth/login or /v1/auth/register. Prefix value with
      "Bearer " (e.g. `Bearer eyJhbG...`).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"